Issue-3499: Move project and network modules to compose/core package
 In order to figure decouple CLI from Core it is necessary
 to understand their cross-dependencies and decouple them
 for further compose/core extraction

Partially implements #3499
