+++ compose/cli/command.py
@@ -11,7 +11,7 @@ from . import verbose_proxy
 from .. import config
 from ..common.const import API_VERSIONS
 from ..config.environment import Environment
-from ..project import Project
+from ..core.project import Project
 from .docker_client import docker_client
 from .docker_client import tls_config_from_options
 from .utils import get_version_info
