+++ tests/unit/project_test.py
@@ -12,7 +12,7 @@ from compose.config.config import Config
 from compose.config.types import VolumeFromSpec
 from compose.core import dockerclient as dc
 from compose.core.container import Container
-from compose.project import Project
+from compose.core.project import Project
 from compose.service import ImageType
 from compose.service import Service
